requirements.txt
google-generativeai
scikit-learn
numpy
flask
flask-session